# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DESIGN.Studio API — print-first AI design studio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, in_progress, ready, archived]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/history:
    get:
      operationId: getProjectHistory
      tags: [projects]
      summary: Get project history timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
  /projects/{id}/backup:
    post:
      operationId: backupProject
      tags: [projects]
      summary: Backup project to GitHub
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Backup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupResult"

  # ── Assets ─────────────────────────────────────────────────────────────────
...
  /tech-packs:
    get:
      operationId: listTechPacks
      tags: [techPack]
      summary: List tech packs
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tech packs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechPack"
    post:
      operationId: createTechPack
      tags: [techPack]
      summary: Generate a tech pack PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechPackInput"
      responses:
        "201":
          description: Generated tech pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechPack"

  # ── Manufacturing ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: ["string", "null"]
        brief:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, in_progress, ready, archived]
        coverAssetUrl:
          type: ["string", "null"]
        colorPalette:
          type: array
          items:
            type: string
        printMethod:
          type: ["string", "null"]
        estimatedQuantity:
          type: ["integer", "null"]
        githubRepo:
          type: ["string", "null"]
        lastBackupAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        brief:
          type: string
        status:
          type: string
          enum: [draft, in_progress, ready, archived]
        printMethod:
          type: string
        estimatedQuantity:
          type: integer
        colorPalette:
          type: array
          items:
            type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        brief:
          type: string
        status:
          type: string
          enum: [draft, in_progress, ready, archived]
        printMethod:
          type: string
        estimatedQuantity:
          type: integer
        colorPalette:
          type: array
          items:
            type: string
        githubRepo:
          type: string
        coverAssetUrl:
          type: string
    HistoryEntry:
      type: object
      required: [id, projectId, type, description, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        description:
          type: string
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BackupResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        commitUrl:
          type: ["string", "null"]

    # ── Assets ────────────────────────────────────────────────────────────────
...
    TechPack:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: ["string", "null"]
        pdfUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        garmentType:
          type: ["string", "null"]
        printMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TechPackInput:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: integer
        title:
          type: string
        garmentType:
          type: string
        printMethod:
          type: string
        placement:
          type: string
        dimensions:
          type: string
        colorCount:
          type: integer
        notes:
          type: string
        mockupAssetUrl:
          type: string

    # ── Manufacturing ─────────────────────────────────────────────────────────
...